rsx: refactor rsx_utils a bit
- Move obviously standalone things to their own utility files
